+++ src/Trait/EventsTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\ActiveRecord\Trait;
 
-use Closure;
+use ReflectionClass;
 use Yiisoft\ActiveRecord\ActiveQueryInterface;
 use Yiisoft\ActiveRecord\ActiveRecordInterface;
 use Yiisoft\ActiveRecord\Event\AfterCreateQuery;
@@ -22,8 +22,6 @@ use Yiisoft\ActiveRecord\Event\BeforeSave;
 use Yiisoft\ActiveRecord\Event\BeforeUpdate;
 use Yiisoft\ActiveRecord\Event\BeforeUpsert;
 use Yiisoft\ActiveRecord\Event\EventDispatcherProvider;
-
-use function is_string;
 
 /**
  * Trait to implement event dispatching for ActiveRecord.
@@ -84,14 +82,11 @@ trait EventsTrait
         return $this;
     }
 
-    public static function query(ActiveRecordInterface|Closure|string|null $modelClass = null): ActiveQueryInterface
+    public static function query(ActiveRecordInterface|string|null $modelClass = null): ActiveQueryInterface
     {
-        $model = match (true) {
-            $modelClass === null => new static(),
-            is_string($modelClass) => new $modelClass(),
-            $modelClass instanceof ActiveRecordInterface => $modelClass,
-            default => ($modelClass)(),
-        };
+        $model = $modelClass instanceof ActiveRecordInterface
+            ? $modelClass
+            : (new ReflectionClass($modelClass ?? static::class))->newInstanceWithoutConstructor();
 
         $eventDispatcher = EventDispatcherProvider::get($model::class);
         $eventDispatcher->dispatch($event = new BeforeCreateQuery($model));