+++ tests/EventsTraitTest.php
@@ -11,7 +11,6 @@ use Yiisoft\ActiveRecord\Event\BeforeSave;
 use Yiisoft\ActiveRecord\Event\BeforeUpdate;
 use Yiisoft\ActiveRecord\Event\BeforeUpsert;
 use Yiisoft\ActiveRecord\Event\EventDispatcherProvider;
-use Yiisoft\ActiveRecord\Tests\Stubs\ActiveRecord\Category;
 use Yiisoft\ActiveRecord\Tests\Stubs\ActiveRecord\CategoryEventsModel;
 use Yiisoft\Test\Support\EventDispatcher\SimpleEventDispatcher;
 
@@ -108,13 +107,6 @@ abstract class EventsTraitTest extends TestCase
         $query = CategoryEventsModel::query();
 
         $this->assertSame($customQuery, $query);
-    }
-
-    public function testQueryWithClosureModelClass(): void
-    {
-        $query = CategoryEventsModel::query(fn() => new Category());
-
-        $this->assertInstanceOf(Category::class, $query->getModel());
     }
 
     public function testSaveWithEventPrevention(): void